REORDER THE SERVICES IN THE INSTALL SERVICES PLAYBOOK ALPHABETICALLY
Now that we are no longer using multiple tags to implement dependencies
between services there is no longer any need to order the services in
this file according to those dependencies. Instead we will implement
dependencies via multiple runs of the playbook; for example run the
playbook first for the DNS service and then again for any services that
are dependence on the DNS service. Also removed some comments that made
reference to the old way of controlling dependencies and so are no
longer applicable.