RENAME WORDPRESS SITE PLAYBOOKS TO HAVE A COMMON PREFIX
I am using the prefix "wp-" for all the playbook files that implement
WordPress site actions so that they can be comprehended as an
operational group more easily. This will make any references to the file
names in the headers incorrect for now but I am committing as is so that
the version history is clearer.