ADD THE DNS API ROLE TO WORDPRESS HOSTS
I have an upgrade workflow which consists of copying a WordPress site to
another host for private testing on the new host before changing the DNS
zone for the project to make the copy on the new host live. This entails
installing a certificate for the WordPress site on the new host before
the DNS record for ths site points to there. We can't meet the certbot
HTTP challenge at that point and so must use the certbot DNS challenge
instead. This requires that WordPress hosts receive the dns_api role.